refactor(scripts/lib): extract move detection and deep-merge helpers
Move collectFeatures/tokenizeLeaf/getAt/detectMoves and deepMerge out
of scripts/diff-flat.js into scripts/lib/detect-moves.js and
scripts/lib/deep-merge.js so other scripts can reuse them. No
behaviour change.
